+++ src/main/kotlin/sparta/nbcamp/wachu/domain/wine/repository/WineQueryDslRepositoryImpl.kt
@@ -1,0 +1,58 @@
+package sparta.nbcamp.wachu.domain.wine.repository
+
+import com.querydsl.core.BooleanBuilder
+import com.querydsl.core.types.Expression
+import com.querydsl.core.types.Order
+import com.querydsl.core.types.OrderSpecifier
+import com.querydsl.core.types.dsl.EntityPathBase
+import com.querydsl.core.types.dsl.PathBuilder
+import org.springframework.data.domain.Page
+import org.springframework.data.domain.PageImpl
+import org.springframework.data.domain.Pageable
+import org.springframework.stereotype.Repository
+import sparta.nbcamp.wachu.domain.wine.entity.QWine
+import sparta.nbcamp.wachu.domain.wine.entity.Wine
+import sparta.nbcamp.wachu.infra.querydsl.QueryDslSupport
+
+@Repository
+class WineQueryDslRepositoryImpl : WineQueryDslRepository, QueryDslSupport() {
+
+    private val wine = QWine.wine
+    override fun searchWines(query: String, pageable: Pageable): Page<Wine> {
+        val whereClause = BooleanBuilder()
+
+        whereClause.and(
+            wine.name.containsIgnoreCase(query)
+                .or(wine.country.containsIgnoreCase(query))
+                .or(wine.region.containsIgnoreCase(query))
+        )
+
+        val totalCount = queryFactory.select(wine.count())
+            .from(wine)
+            .where(whereClause)
+            .fetchOne() ?: 0L
+
+        val contents = queryFactory.selectFrom(wine)
+            .where(whereClause)
+            .offset(pageable.offset)
+            .limit(pageable.pageSize.toLong())
+            .orderBy(*getOrderSpecifier(pageable, wine))
+            .fetch()
+
+        return PageImpl(contents, pageable, totalCount)
+    }
+
+    private fun getOrderSpecifier(
+        pageable: Pageable,
+        path: EntityPathBase<*>
+    ): Array<OrderSpecifier<*>> {
+        val pathBuilder = PathBuilder(path.type, path.metadata)
+
+        return pageable.sort.toList().map { order ->
+            OrderSpecifier(
+                if (order.isAscending) Order.ASC else Order.DESC,
+                pathBuilder.get(order.property) as Expression<Comparable<*>>
+            )
+        }.toTypedArray()
+    }
+}
